# docs/swagger.yaml
basePath: /
definitions:
...
  models.Student:
    properties:
      class:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
...
host: localhost:443
info:
  contact: {}
  description: This is a REST API for managing Teachers, Students and Executives
  title: REST API Documentation
  version: "1.0"
paths:
  /allstudents/:
    delete:
      description: Delete all students from the system
      responses:
        "200":
          description: All students deleted successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete all students
      tags:
      - Students
...
  /students/:
    patch:
      consumes:
      - application/json
      description: Partially update multiple students
      parameters:
      - description: Student updates
        in: body
        name: updates
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Students patched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Patch multiple students
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Create one or multiple new students
      parameters:
      - description: Student data
        in: body
        name: students
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Student'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Students created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Add new students
      tags:
      - Students
  /students/{id}:
    delete:
      description: Delete a specific student by their ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Student deleted successfully
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete student
      tags:
      - Students
    get:
      description: Get a specific student by their ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get student by ID
      tags:
      - Students
    patch:
      consumes:
      - application/json
      description: Partially update a specific student by ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student updates
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Student patched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Patch student
      tags:
      - Students
    put:
      consumes:
      - application/json
      description: Update a specific student by their ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student data to update
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: Student updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update student
      tags:
      - Students
...
schemes:
- https
securityDefinitions:
  Bearer:
    description: JWT Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"